Thrown out the line-height-fit effort.
This is a very nice thing to have, but it's pretty "expensive" to set up, plus
it makes the stylesheets appear to have too much "magic". Its intensive need
for attention held back the development of other features.

The mixin is being left in place, just in case there is a specific project
where it's deemed worthy of implementation.